check for more than one alternative of /bin/true
Until now code only checked for /bin/true; this is a problem when source
OS and destionation OS differ. E.g. when dest OS is Linux and source OS
is FreeBSD, repmgr would only check for /bin/true (because it has been
compiled on Linux) and therefore fail. Now it checks both.