fixes a bug in pg data directory creation
Beginning with the last commit fixing the exit() during a failure in
creating the pg data directory did not fail with a negative error
code. Now we added a new flag `flag_fatal`; when set, we skip everything
following the stop_backup label but the pg_stop_backup() and return a
negative error code.