Fix -D/--data-dir's long option
This was erroneously rendered as --dest-dir. This bugfeature seems
to have been around for a long time; however as the only way anyone
could know of the existence of --dest-dir is by reading the source
code, we can safely remove it.

Pointed out by Github user Jehan-Guillaume (ioguix) de Rorthais.