Fixes #312
In order to compute the location of the base backups, we were
retrieving the value of `backup_directory` for the corresponding
Barman server, and then appending `/base`.

This is Barman's default location, but it can be overridden by
specifying an explicit setting of `basebackups_directory`, in which
case repmgr would be unable to remove the prefix from the paths
retrieved from Barman, failing with an error message

    Unexpected output from "barman list-files": (...)

As we do not use `backup_directory` anywhere else, the solution is
simple, i.e. retrieve `basebackups_directory` directly.

The problem was originally reported by David Dumas as a Barman issue:

  https://github.com/2ndquadrant-it/barman/issues/107